Add DTOs and improve separation of concerns

+++ WildlifeTracker/Data/Models/Animal.cs
@@ -6,7 +6,7 @@ namespace WildlifeTracker.Data.Models
         public required string Species { get; set; }
         public int Age { get; set; }
         public double Weight { get; set; }
-        public int Height { get; set; }
+        public double Height { get; set; }
 
         public ICollection<Sighting> Sightings { get; set; } = new HashSet<Sighting>();
     }
